requirements.txt
fastapi
uvicorn
httpx
python-multipart
pydantic
pydantic-settings
langchain
langchain-core
langchain-deepseek
pymupdf
pillow
sympy
pandas
openpyxl
